requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aio-pika==9.3.1
    # via omotes-sdk-python
aiormq==6.7.7
    # via aio-pika
amqp==5.2.0
    # via kombu
billiard==4.2.0
    # via celery
celery==5.3.6
    # via omotes-sdk-python
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
coloredlogs==15.0.1
    # via
    #   omotes-simulator-core
    #   simulator-worker (pyproject.toml)
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclass-wizard==0.22.2
    # via omotes-simulator-core
deepdiff==6.7.1
    # via pandapower
fonttools==4.50.0
    # via matplotlib
future-fstrings==1.2.0
    # via pyecore
humanfriendly==10.0
    # via coloredlogs
idna==3.6
    # via
    #   requests
    #   yarl
influxdb==5.3.1
    # via omotes-simulator-core
kiwisolver==1.4.5
    # via matplotlib
kombu==5.3.5
    # via celery
lxml==5.1.0
    # via pyecore
mako==1.3.2
    # via pdoc3
markdown==3.6
    # via pdoc3
markupsafe==2.1.5
    # via mako
matplotlib==3.8.3
    # via pandapipes
msgpack==1.0.8
    # via influxdb
multidict==6.0.5
    # via yarl
networkx==3.2.1
    # via pandapower
numpy==1.26.0
    # via
    #   contourpy
    #   matplotlib
    #   omotes-simulator-core
    #   pandapower
    #   pandas
    #   scipy
    #   shapely
omotes-sdk-protocol==0.0.8
    # via omotes-sdk-python
omotes-sdk-python==0.0.14
    # via simulator-worker (pyproject.toml)
omotes-simulator-core==0.0.8
    # via simulator-worker (pyproject.toml)
ordered-set==4.1.0
    # via
    #   deepdiff
    #   pyecore
packaging==24.0
    # via
    #   matplotlib
    #   pandapower
pamqp==3.2.1
    # via aiormq
pandapipes==0.8.4
    # via omotes-simulator-core
pandapower==2.13.1
    # via pandapipes
pandas==2.1.1
    # via
    #   omotes-simulator-core
    #   pandapower
    #   simulator-worker (pyproject.toml)
pdoc3==0.10.0
    # via streamcapture
pillow==10.2.0
    # via matplotlib
prompt-toolkit==3.0.43
    # via click-repl
protobuf==4.25.3
    # via omotes-sdk-protocol
pyecore==0.12.1
    # via pyesdl
pyesdl==24.2
    # via
    #   omotes-simulator-core
    #   simulator-worker (pyproject.toml)
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   celery
    #   influxdb
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via simulator-worker (pyproject.toml)
pytz==2024.1
    # via
    #   influxdb
    #   pandas
requests==2.31.0
    # via influxdb
restrictedpython==7.2a1.dev0
    # via pyecore
scipy==1.12.0
    # via pandapower
shapely==2.0.3
    # via pandapipes
six==1.16.0
    # via
    #   influxdb
    #   python-dateutil
streamcapture==1.2.2
    # via omotes-sdk-python
tqdm==4.66.2
    # via pandapower
tzdata==2024.1
    # via
    #   celery
    #   pandas
urllib3==2.2.1
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.9.4
    # via
    #   aio-pika
    #   aiormq

# The following packages are considered to be unsafe in a requirements file:
# setuptools
